Krish: Major and Final Week 2 Update (v2.9)
This is a big one. The reworked popup system has been completed in this version, the ProblemPanel has been created (with MultipleChoiceQuestion and ShortAnswerQuestion), new utility methods have been added to GameData regarding String parsing, a new text file has been created for reading in problems, and some minor updates have been made in terms of documentation and just code in general.

Currently, a lot of classes are still incomplete, but we can finish them next week.

This is definitely a great place to end off our second week, with a ton of backend stuff completed. Hopefully the third week will be incredibly productive for us!