Update v3.10
A major update to gameplay, as questions are now able to be answered. As of now, the intermission panel will take you straight to a problem, but in the future, we should integrate it into the actual game.

Problems can either be multiple choice or short answer, and the user can respond to these questions. Once a response has been complete, they can press the submit button, or if they opened a question on accident, they can return back to the intermission panel. When a question is submitted, the answer is reviewed, and a popup displaying the answer and explanations will appear.